+++ ReleaseNotes/release_notes.md
@@ -15,6 +15,8 @@
 
 ## Bug fixes:
 
+  1. fix `Ostap::Utils::hahs_histo` and `Ostap::Utils::hahs_axis`
+
 # v1.9.7.6
 
 ## New features: